Work around the lack of QUiLoader in PyQt
